Fix pointer type issues for macOS compatibility
- Fixed Swift pointer assignments to match system zlib definition
- System zlib on macOS defines next_in as mutable (Bytef*) not const
- Updated all Swift code to cast const pointers to mutable when assigning to next_in
- Simplified header configuration to always use system zlib on macOS
- Moved zlib_simple.h to private directory to avoid module conflicts
- Added z_const definition to ensure proper const qualifier interpretation
- All 262 tests now pass on macOS

This resolves the pointer type mismatch errors that were preventing
the build from succeeding on macOS with system zlib.

+++ Sources/CZLib/zlib_shim.c
@@ -262,7 +262,7 @@ static unsigned int debug_in_wrapper(void* desc, unsigned char** buf) {
     swift_inflateback_context_t* context = (swift_inflateback_context_t*)desc;
     unsigned int result = 0;
     if (context && context->swift_in_func) {
-        result = context->swift_in_func(context->swift_context, buf, NULL);
+        result = (unsigned int)context->swift_in_func(context->swift_context, buf, NULL);
 #if ZLIB_DEBUG
         if (result == 0) {
             printf("[C shim] input callback returned 0 (no data)\n");
